[plugin.video.vrt.nu] 1.7.0
This is an urgent release as due to changes to the VRT.NU website, our
users are no longer able to access program listings and categories.

The new code actually uses more API calls so that this the probablity of
this happening in the future is less likely.

This code also (optionally) requires **script.module.pysocks** and
**script.module.win_inet_pton** for Proxy and SOCKS support.